Preliminary work to implement SCADA controls. There is a new subclass in controls.py that can implement SCADA control. This type of controls polls the SCADA for the status of an actuator and sets that value into the DB. This centralizes the control logic into the SCADA instead of the PLCs
The genericSCADA has the responsibility of calculating the actuator values.
The actuator values are handled in the same cache way of other values and are constantly being available via a thread.

A function called "update_actuator_values" can update this cache. This function could receive a list of dictionaries with keys=tag values=calculated values. For now is just alternating between 1-0 for all the available pumps